fix skiplist error

+++ src/util/inc/tskiplist.h
@@ -54,7 +54,7 @@ typedef struct SSkipListNode {
 #define SL_GET_NODE_DATA(n) ((char*)(n) + SL_NODE_HEADER_SIZE((n)->level))
 #define SL_GET_NODE_KEY(s, n) ((s)->keyFn(SL_GET_NODE_DATA(n)))
 
-#define SL_GET_NODE_LEVEL(n) *(int32_t *)((n))
+#define SL_GET_NODE_LEVEL(n) *(uint8_t *)((n))
 
 /*
  * @version 0.3
